backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
black==24.8.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
flake8==7.1.1
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
